mysite/requirements.txt
Django==3.0
psycopg2-binary==2.8.4
gunicorn==20.0.4
Pillow==7.1.2
djangorestframework
django-rest-swagger